Refactor diagnostics
