+++ examples/09_bans/deny.toml
@@ -10,8 +10,6 @@ deny = [
     { name = "openssl-sys" },
 ]
 skip = [
-    # rustls uses an old version of base64
-    { name = "base64", version = "0.10" },
     # miow unfortunately still uses the ancient 0.2 version of winapi
     { name = "winapi", version = "=0.2.8" },
 ]
