+++ src/advisories/mod.rs
@@ -1,4 +1,5 @@
 pub mod cfg;
+mod diags;
 pub mod fix;
 mod helpers;
 
@@ -35,14 +36,11 @@ pub fn check<R>(
     advisory_dbs: &DbSet,
     lockfile: PrunedLockfile,
     audit_compatible_reporter: Option<R>,
-    sender: crossbeam::channel::Sender<diag::Pack>,
+    mut sink: diag::ErrorSink,
 ) where
     R: AuditReporter,
 {
-    use rustsec::{
-        advisory::{informational::Informational, metadata::Metadata},
-        package::Package,
-    };
+    use rustsec::{advisory::metadata::Metadata, package::Package};
 
     let emit_audit_compatible_reports = audit_compatible_reporter.is_some();
 
@@ -70,119 +68,26 @@ pub fn check<R>(
     use bitvec::prelude::*;
     let mut ignore_hits = bitvec![0; ctx.cfg.ignore.len()];
 
-    let mut make_diag = |pkg: &Package, advisory: &Metadata| -> diag::Pack {
-        match krate_for_pkg(&ctx.krates, pkg) {
-            Some((i, krate)) => {
-                let id = &advisory.id;
+    let mut send_diag = |pkg: &Package, advisory: &Metadata| match krate_for_pkg(&ctx.krates, pkg) {
+        Some((i, krate)) => {
+            let diag = ctx.diag_for_advisory(krate, i, advisory, |index| {
+                ignore_hits.as_mut_bitslice().set(index, true)
+            });
 
-                let (severity, message) = {
-                    let (lint_level, msg) = match &advisory.informational {
-                        // Everything that isn't an informational advisory is a vulnerability
-                        None => (
-                            ctx.cfg.vulnerability,
-                            "security vulnerability detected".to_owned(),
-                        ),
-                        Some(info) => match info {
-                            // Security notices for a crate which are published on https://rustsec.org
-                            // but don't represent a vulnerability in a crate itself.
-                            Informational::Notice => {
-                                (ctx.cfg.notice, "notice advisory detected".to_owned())
-                            }
-                            // Crate is unmaintained / abandoned
-                            Informational::Unmaintained => (
-                                ctx.cfg.unmaintained,
-                                "unmaintained advisory detected".to_owned(),
-                            ),
-                            Informational::Unsound => {
-                                (ctx.cfg.unsound, "unsound advisory detected".to_owned())
-                            }
-                            // Other types of informational advisories: left open-ended to add
-                            // more of them in the future.
-                            Informational::Other(_) => {
-                                unreachable!("rustsec only returns these if we ask, and there are none at the moment to ask for");
-                            }
-                            _ => unreachable!("unknown advisory type encountered"),
-                        },
-                    };
-
-                    // Ok, we found a crate whose version lies within the range of an
-                    // advisory, but the user might have decided to ignore it
-                    // for "reasons", but in that case we still emit it to the log
-                    // so it doesn't just disappear into the aether
-                    let lint_level =
-                        if let Ok(index) = ctx.cfg.ignore.binary_search_by(|i| i.value.cmp(id)) {
-                            ignore_hits.as_mut_bitslice().set(index, true);
-                            LintLevel::Allow
-                        } else if let Some(severity_threshold) = ctx.cfg.severity_threshold {
-                            if let Some(advisory_severity) =
-                                advisory.cvss.as_ref().map(|cvss| cvss.severity())
-                            {
-                                if advisory_severity < severity_threshold {
-                                    LintLevel::Allow
-                                } else {
-                                    lint_level
-                                }
-                            } else {
-                                lint_level
-                            }
-                        } else {
-                            lint_level
-                        };
-
-                    (
-                        match lint_level {
-                            LintLevel::Warn => Severity::Warning,
-                            LintLevel::Deny => Severity::Error,
-                            LintLevel::Allow => Severity::Help,
-                        },
-                        msg,
-                    )
-                };
-
-                let notes = {
-                    let mut n = Vec::new();
-
-                    n.push(advisory.description.clone());
-
-                    if let Some(ref url) = advisory.url {
-                        n.push(format!("URL: {}", url));
-                    }
-
-                    n
-                };
-
-                let mut pack = diag::Pack::with_kid(Check::Advisories, krate.id.clone());
-                let diag = pack.push(
-                    Diagnostic::new(severity)
-                        .with_message(advisory.title.clone())
-                        .with_labels(vec![ctx.label_for_span(i.index(), message)])
-                        .with_code(id.as_str().to_owned())
-                        .with_notes(notes),
-                );
-
-                if ctx.serialize_extra {
-                    diag.extra = serde_json::to_value(&advisory)
-                        .ok()
-                        .map(|v| ("advisory", v));
-                }
-
-                pack
-            }
-            None => {
-                unreachable!(
-                    "the advisory database report contained an advisory 
+            sink.push(diag);
+        }
+        None => {
+            unreachable!(
+                "the advisory database report contained an advisory 
                     that somehow matched a crate we don't know about:\n{:#?}",
-                    advisory
-                );
-            }
+                advisory
+            );
         }
     };
 
     // Emit diagnostics for any vulnerabilities that were found
     for vuln in &report.vulnerabilities {
-        sender
-            .send(make_diag(&vuln.package, &vuln.advisory))
-            .unwrap();
+        send_diag(&vuln.package, &vuln.advisory);
     }
 
     // Emit diagnostics for informational advisories for crates, including unmaintained and unsound
@@ -190,46 +95,26 @@ pub fn check<R>(
         .iter_warnings()
         .filter_map(|(_, wi)| wi.advisory.as_ref().map(|wia| (wi, wia)))
     {
-        sender.send(make_diag(&warning.package, &advisory)).unwrap();
+        send_diag(&warning.package, &advisory);
     }
 
     match yanked {
         Ok(yanked) => {
             for pkg in yanked {
-                let diag = match krate_for_pkg(&ctx.krates, &pkg) {
+                match krate_for_pkg(&ctx.krates, &pkg) {
                     Some((ind, krate)) => {
-                        let mut pack = diag::Pack::with_kid(Check::Advisories, krate.id.clone());
-                        pack.push(
-                            Diagnostic::new(match ctx.cfg.yanked.value {
-                                LintLevel::Allow => Severity::Help,
-                                LintLevel::Deny => Severity::Error,
-                                LintLevel::Warn => Severity::Warning,
-                            })
-                            .with_message("detected yanked crate")
-                            .with_labels(vec![ctx.label_for_span(ind.index(), "yanked version")]),
-                        );
-
-                        pack
+                        sink.push(ctx.diag_for_yanked(krate, ind));
                     }
                     None => unreachable!(
                         "the advisory database warned about yanked crate that we don't have: {:#?}",
                         pkg
                     ),
                 };
-
-                sender.send(diag).unwrap();
             }
         }
         Err(e) => {
             if ctx.cfg.yanked.value != LintLevel::Allow {
-                let mut diag = diag::Pack::new(Check::Advisories);
-                diag.push(
-                    Diagnostic::warning()
-                        .with_message(format!("unable to check for yanked crates: {}", e))
-                        .with_labels(vec![Label::primary(ctx.cfg.file_id, ctx.cfg.yanked.span)
-                            .with_message("lint level defined here")]),
-                );
-                sender.send(diag).unwrap();
+                sink.push(ctx.diag_for_index_failure(e));
             }
         }
     }
@@ -260,25 +145,14 @@ pub fn check<R>(
 
     // Check for advisory identifers that were set to be ignored, but
     // were not actually encountered, for cases where a crate, or specific
-    // verison of that crate, has been removed or replaced and the advisory
-    // no longer applies to it so that users can cleanup their configuration
+    // version of that crate, has been removed or replaced and the advisory
+    // no longer applies to it, so that users can cleanup their configuration
     for ignore in ignore_hits
         .into_iter()
-        .zip(ctx.cfg.ignore.into_iter())
+        .zip(ctx.cfg.ignore.iter())
         .filter_map(|(hit, ignore)| if !hit { Some(ignore) } else { None })
     {
-        sender
-            .send(
-                (
-                    Check::Advisories,
-                    Diagnostic::warning()
-                        .with_message("advisory was not encountered")
-                        .with_labels(vec![Label::primary(ctx.cfg.file_id, ignore.span)
-                            .with_message("no crate matched advisory criteria")]),
-                )
-                    .into(),
-            )
-            .unwrap();
+        sink.push(ctx.diag_for_advisory_not_encountered(ignore));
     }
 
     if let Some(mut reporter) = audit_compatible_reporter {
